SequentialExecutor support for pmap, explicitly set module outputs (#7)
This PR also includes a change I made to Module: I've removed the code that automatically sets outputs to be the sinks of the module, and instead added an explicit method that sets the outputs of a module to a given list of values. If we want the old behaviour, we can call a set_outputs_auto method that sets outputs automatically to the sinks. This function is also called by a new method called finalize that performs verifications and sets outputs if none are set already.

* Initial pmap implementation in SequentialExecutor

* Add some tests (and fix linting issues)

* module: outputs should be explicitly set

* Fix tests

* Add some type hints and comments

* Add some type hints and comments

* Add infer_shapes to Module.finalize